[CoreLocation] Add support for Xcode 12 beta 1. (#9001)
A number of APIs added and deprecated in the same release. We will see
that is that about.

The status property move to be a instance property in CLLocationManager,
we expose those and add a deprecation warning. That needs to be ignored in xtro due to issue https://github.com/xamarin/xamarin-macios/issues/9026


Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>