[monotouch-test] Fix broken UrlRequestTest.Mutability_30744 test. (#9096)
* [monotouch-test] Fix broken UrlRequestTest.Mutability_30744 test.

This test is supposed to assert that setting a header on an immutable
dictionary throws an exception.

The problem is that the tested code always throws an exception: either when
setting the header, or when asserting right after because no exception was
thrown, effectively always passing.

This showed up with current versions of NUnit[Lite], which keeps track of
assertion failures, and fails a test even if those assertion
failures/exceptions are caught in exception handlers.

So remove the whole exception verification logic, and just try to set the
value. We'll see if the test start failing somewhere (I've already tested both
on device and simulator, and no problems found so far).

* Remove a simulator condition.

It works on device for me (iPhone 7 with iOS 11.4.1)... maybe Apple changed
something?