[tests] The upstream NUnitLine isn't linker safe, so add an xml file to make sure linking doesn't remove too much.
In particular NUnit uses reflection to get a private method, and the linker removes
the corresponding private method:

https://github.com/nunit/nunit/blob/1c680b4dc8ad3c6caf6563f434e2c98caac2568d/src/NUnitFramework/framework/Internal/TestExecutionContext.cs#L552

So add an xml definition to keep this private method, and modify project files to
pass the xml definition to mtouch and mmp.

Some care needs to be taken to make sure xharness is still able to clone these project
files.