[DevOps] Report the correct bot used. (#9250)
* [DevOps] Report the correct bot used.

The tests are not executed in the same agent that sets the github
status, for that reason, the github comment is not giving the correct
bot name, but the name of the bot that executed the commit message and
not the tests.

Add an output var in the runTests step to set the bot name and pass it
to the comment job in a different bot.

* Update tools/devops/device-tests/scripts/GitHub.psm1

Co-authored-by: Whitney Schmidt <whschm@microsoft.com>

Co-authored-by: Whitney Schmidt <whschm@microsoft.com>