[iOSurface] Remove warning from the generator. (#9357)
The following warning was being raised:

```
iosurface.cs(202,4): warning CS0657: 'return' is not a valid attribute location for this declaration. Valid attribute locations for this declaration are 'property'. All attributes in this block will be ignored.
```

Moved the attr to the correct location and fixed a missing using with
the MaybeNull attrs.