[tests][intro] Split attributes typos from API typos tests
The latter requires the spellchecker which varies by OS versions, so we
only run it on the latest OS version (and simulator, except macOS).

The former are some internal rules and can be run on every commit and
avoid finding issues late in a release cycle.

Also changed to
- process members even if a type is obsoleted
- process the properties and events on types

Note: The PR doe snot include any fixes (to typos) since it would
make the xcode12 merge harder. They will be fixed later