[tests][xtro] Add enum member/values checks. Fix #7527
Figure out if
* we're missing enum values (easy to workaround, but annoying for developers)
* we expose enum values that are not defined natively (potential bugs)

reference: https://github.com/xamarin/xamarin-macios/issues/7527

Draft for the moment. Fixes in the platform assemblies will be made
later, once `xcode12` is merged into `main`. IOW there will be errors
from xtro and that's good to check them to see if there's bugs in the
logic, ways to reduce noise (or just use the `.ignore` files for things
we can't fix)