[Harness] Fix device tests by ensure that the lambdas are correctly chained.
The launch continuation was not being used correctly since it was using
a Task<bool> on a Task. That is possible because class Task<T> : Task
but was giving runtime issues.

The problem with this case is that, since the launch callback was not
called, although the test were executed, it would set the test as a
timeout. With the change, the launch callback is always executed
allowing the results to be reported correctly.

fixes https://github.com/xamarin/maccore/issues/2316
fixes https://github.com/xamarin/maccore/issues/2185